requirements.txt
fastapi
uvicorn
pydantic
pytest
faker
httpx
psycopg2-binary
python-dotenv